+++ R/splits.R
@@ -42,7 +42,7 @@ get_splits <- function(ticker = "AAPL", exchange = "US",
 
   content <- query_api(url)
 
-  if (content == "[]") {
+  if (is_empty_body(content)) {
     cli::cli_alert_danger("cant find split data for {ticker}|{exchange}")
 
     df_split <- dplyr::tibble()